Code improvement
