+++ icmp/netwicmp_table.h
@@ -1,0 +1,8 @@
+#ifndef LIGHT_NETW_NETWICMP_TABLE_H
+#define LIGHT_NETW_NETWICMP_TABLE_H
+
+#include "netwicmp.h"
+
+void print_ping_table(size_t packets_c, icmp_transmis packets[packets_c]);
+
+#endif //LIGHT_NETW_NETWICMP_TABLE_H
